+++ src/app/(group)/blog/page.tsx
@@ -5,6 +5,7 @@ import { sanityFetch } from "../../../../sanity/lib/fetch";
 import { postsQuery } from "../../../../sanity/lib/queries";
 import { SanityDocument } from "next-sanity";
 import { Metadata } from "next";
+import Newsletter from "@/components/newsletter";
 
 export const metadata: Metadata = {
     title: 'Blog | Book Events Africa',
@@ -17,7 +18,7 @@ async function Blog() {
     const posts = await sanityFetch<SanityDocument>({ query: postsQuery });
     return <>
         <Navbar />
-
+        <Newsletter tag={'blog'} />
         <section className="px-[1rem] lg:px-[14rem] bg-[#F1F2F6] py-[20px] lg:py-[56px]">
             <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 px-0 lg:px-5 gap-[29px]">
                 {posts.map((post: any, i: number) => {
